requirements.txt
APScheduler==3.9.1
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
bcrypt==3.2.0
Django==2.1.15
django-mssql==1.8
django-pymssql==1.7.1
django-pyodbc-azure==2.1.0.0
pipenv==2022.9.24
pymssql==2.2.5
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dotenv==0.21.0
Scrapy==2.6.3
scrapy-autounit==0.0.32
scrapy-do==0.5.0
scrapyd==1.3.0
scrapyd-client @ git+https://github.com/scrapy/scrapyd-client.git@7b737c04420a698543b9866dc0adf02d9c30fc88
scrapydweb @ git+https://github.com/my8100/scrapydweb.git@7a3b81dba2cba4279c9465064a693bb277ac20e9
SecretStorage==3.3.1
service-identity==21.1.0
six==1.16.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
whitenoise
numpy==1.23.3

